refactor: move admin views related components
